make GetPasswd take a terminal file descriptor
There are times when we need user input from the terminal, even though
our stdin is piped from something else. (For example, reading an
encrypted file over stdin, and asking the user for the password.) Make
GetPasswd* take an explicit file descriptor for the terminal, so that
the caller has the option of using /dev/tty.